feat: Controller with the endpoint of the packet states and packet information
