+++ routes/index.js
@@ -1,17 +1,16 @@
 import express from "express";
 import { Login, Register, Logout } from "../controllers/client.js";
-import { Package, GetPackage  } from "../controllers/package.js";
+import { Package, GetPackage,GetPackageStates  } from "../controllers/package.js";
 import { refreshToken } from "../controllers/refreshToken.js";
 import { EndUsers } from "../controllers/endUsers.js";
-// import { RegisterStaff } from "../controllers/staff.js";
-import { GetPackageStates } from "../controllers/statesPakage.js";
+import { RegisterStaff } from "../controllers/staff.js";
 
 
 const router = express.Router();
 
 // POST
 router.post("/register", Register);
-// router.post("/registerStaff", RegisterStaff);
+router.post("/registerStaff", RegisterStaff);
 router.post("/login", Login);
 router.post("/package", Package);
 router.post("/addEndUsers", EndUsers);
